Remove test_gradients
